SLEEPMODEL/requirements.txt
fastapi
uvicorn
pandas
numpy
joblib
scikit-learn
pydantic
google-generativeai
python-dotenv